Split profile folder into TeamProfile and PlayerProfile folders
This commit separates the profile folder into TeamProfile and PlayerProfile. This is done to improve code organization of future commits.
